chore: update proto files
ENG-145
