+++ indykite_sdk_node/src/grpc/indykite/identity/v1beta2/model.ts
@@ -27,6 +27,7 @@ import { MessageType } from "@protobuf-ts/runtime";
 import { Struct } from "../../../google/protobuf/struct";
 import { MapValue } from "../../objects/v1beta1/struct";
 import { Value } from "../../../google/protobuf/struct";
+import { PropertyFilter } from "./attributes";
 import { Property } from "./attributes";
 import { Timestamp } from "../../../google/protobuf/timestamp";
 /**
@@ -85,6 +86,35 @@ export interface DigitalEntity {
      * @generated from protobuf field: repeated indykite.identity.v1beta2.Property properties = 3;
      */
     properties: Property[];
+}
+/**
+ * @generated from protobuf message indykite.identity.v1beta2.DigitalTwinIdentifier
+ */
+export interface DigitalTwinIdentifier {
+    /**
+     * @generated from protobuf oneof: filter
+     */
+    filter: {
+        oneofKind: "digitalTwin";
+        /**
+         * @generated from protobuf field: indykite.identity.v1beta2.DigitalTwin digital_twin = 1;
+         */
+        digitalTwin: DigitalTwin;
+    } | {
+        oneofKind: "propertyFilter";
+        /**
+         * @generated from protobuf field: indykite.identity.v1beta2.PropertyFilter property_filter = 2;
+         */
+        propertyFilter: PropertyFilter;
+    } | {
+        oneofKind: "accessToken";
+        /**
+         * @generated from protobuf field: string access_token = 3;
+         */
+        accessToken: string;
+    } | {
+        oneofKind: undefined;
+    };
 }
 /**
  * ErrorMessage see [Problem Details for HTTP APIs](https://tools.ietf.org/html/rfc7807)
@@ -977,6 +1007,76 @@ class DigitalEntity$Type extends MessageType<DigitalEntity> {
  * @generated MessageType for protobuf message indykite.identity.v1beta2.DigitalEntity
  */
 export const DigitalEntity = new DigitalEntity$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class DigitalTwinIdentifier$Type extends MessageType<DigitalTwinIdentifier> {
+    constructor() {
+        super("indykite.identity.v1beta2.DigitalTwinIdentifier", [
+            { no: 1, name: "digital_twin", kind: "message", oneof: "filter", T: () => DigitalTwin, options: { "validate.rules": { message: { required: true } } } },
+            { no: 2, name: "property_filter", kind: "message", oneof: "filter", T: () => PropertyFilter, options: { "validate.rules": { message: { required: true } } } },
+            { no: 3, name: "access_token", kind: "scalar", oneof: "filter", T: 9 /*ScalarType.STRING*/, options: { "validate.rules": { string: { minLen: "20" } } } }
+        ]);
+    }
+    create(value?: PartialMessage<DigitalTwinIdentifier>): DigitalTwinIdentifier {
+        const message = { filter: { oneofKind: undefined } };
+        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
+        if (value !== undefined)
+            reflectionMergePartial<DigitalTwinIdentifier>(this, message, value);
+        return message;
+    }
+    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: DigitalTwinIdentifier): DigitalTwinIdentifier {
+        let message = target ?? this.create(), end = reader.pos + length;
+        while (reader.pos < end) {
+            let [fieldNo, wireType] = reader.tag();
+            switch (fieldNo) {
+                case /* indykite.identity.v1beta2.DigitalTwin digital_twin */ 1:
+                    message.filter = {
+                        oneofKind: "digitalTwin",
+                        digitalTwin: DigitalTwin.internalBinaryRead(reader, reader.uint32(), options, (message.filter as any).digitalTwin)
+                    };
+                    break;
+                case /* indykite.identity.v1beta2.PropertyFilter property_filter */ 2:
+                    message.filter = {
+                        oneofKind: "propertyFilter",
+                        propertyFilter: PropertyFilter.internalBinaryRead(reader, reader.uint32(), options, (message.filter as any).propertyFilter)
+                    };
+                    break;
+                case /* string access_token */ 3:
+                    message.filter = {
+                        oneofKind: "accessToken",
+                        accessToken: reader.string()
+                    };
+                    break;
+                default:
+                    let u = options.readUnknownField;
+                    if (u === "throw")
+                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
+                    let d = reader.skip(wireType);
+                    if (u !== false)
+                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
+            }
+        }
+        return message;
+    }
+    internalBinaryWrite(message: DigitalTwinIdentifier, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
+        /* indykite.identity.v1beta2.DigitalTwin digital_twin = 1; */
+        if (message.filter.oneofKind === "digitalTwin")
+            DigitalTwin.internalBinaryWrite(message.filter.digitalTwin, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
+        /* indykite.identity.v1beta2.PropertyFilter property_filter = 2; */
+        if (message.filter.oneofKind === "propertyFilter")
+            PropertyFilter.internalBinaryWrite(message.filter.propertyFilter, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
+        /* string access_token = 3; */
+        if (message.filter.oneofKind === "accessToken")
+            writer.tag(3, WireType.LengthDelimited).string(message.filter.accessToken);
+        let u = options.writeUnknownFields;
+        if (u !== false)
+            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
+        return writer;
+    }
+}
+/**
+ * @generated MessageType for protobuf message indykite.identity.v1beta2.DigitalTwinIdentifier
+ */
+export const DigitalTwinIdentifier = new DigitalTwinIdentifier$Type();
 // @generated message type with reflection information, may provide speed optimized methods
 class ErrorMessage$Type extends MessageType<ErrorMessage> {
     constructor() {
